// src/types/User.ts
export type UserProps = {
  name: string;
  cnpj: string;
  email: string;
  telephone: string;
  city: string;
  district: string;
  street: string;
  fuelStationNumber: string;
  cep: string;
  flag: string;
  isNetwork: string;
  networkName: string;
  password: string;
  passwordConfirmation: string;
};